+++ libass/ass_cache_template.h
@@ -32,7 +32,7 @@
 #elif defined(CREATE_HASH_FUNCTIONS)
 #undef CREATE_HASH_FUNCTIONS
 #define START(funcname, structname) \
-    static uint32_t funcname##_hash(void *buf, uint32_t hval) \
+    static uint64_t funcname##_hash(void *buf, uint64_t hval) \
     { \
         struct structname *p = buf;
 #define GENERIC(type, member) \
